requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
fastmcp==0.1.0
streamlit==1.28.1
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
aiohttp==3.8.6
google-generativeai==0.3.2
python-dotenv==1.0.0
